# docs/swagger.yaml
basePath: /api/
definitions:
  handler.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  models.CreateTransportInput:
    properties:
      canBeRented:
        type: boolean
      color:
        type: string
      dayPrice:
        minimum: 1
        type: number
      description:
        type: string
      identifier:
        type: string
      latitude:
        maximum: 180
        minimum: -180
        type: number
      longitude:
        maximum: 180
        minimum: -180
        type: number
      minutePrice:
        minimum: 1
        type: number
      model:
        type: string
      transportType:
        type: string
    required:
    - canBeRented
    - color
    - identifier
    - latitude
    - longitude
    - model
    - transportType
    type: object
  models.CreateTransportOutput:
    properties:
      transport:
        $ref: '#/definitions/models.GetTransportOutput'
    type: object
...
  models.GetTransportOutput:
    properties:
      canBeRented:
        type: boolean
      color:
        type: string
      dayPrice:
        type: number
      description:
        type: string
      id:
        type: integer
      identifier:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      minutePrice:
        type: number
      model:
        type: string
      ownerId:
        type: integer
      transportType:
        type: string
    type: object
...
  models.UpdateTransportInput:
    properties:
      canBeRented:
        type: boolean
      color:
        type: string
      dayPrice:
        minimum: 1
        type: number
      description:
        type: string
      identifier:
        type: string
      latitude:
        maximum: 180
        minimum: -180
        type: number
      longitude:
        maximum: 180
        minimum: -180
        type: number
      minutePrice:
        minimum: 1
        type: number
      model:
        type: string
    required:
    - canBeRented
    - color
    - identifier
    - latitude
    - longitude
    - model
    type: object
  models.UpdateTransportOutput:
    properties:
      transport:
        $ref: '#/definitions/models.GetTransportOutput'
    type: object
info:
  contact: {}
  description: API for transportation rental service
  title: SimbirGoAPI
  version: 1.0.0
paths:
...
  /Transport:
    post:
      consumes:
      - application/json
      description: Adding new transportation.
      parameters:
      - description: '-'
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateTransportInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreateTransportOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Create transport
      tags:
      - transport
  /Transport/{id}:
    delete:
      consumes:
      - application/json
      description: Deleting vehicles by id.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Delete transport
      tags:
      - transport
    get:
      consumes:
      - application/json
      description: Getting information about transport by id.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetTransportOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      summary: Get transport
      tags:
      - transport
    put:
      consumes:
      - application/json
      description: Update transport by id.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: string
      - description: '-'
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTransportInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UpdateTransportOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Update transport
      tags:
      - transport
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"